v4l2-renderer: Rendering optimization.
Any surface that are under opaque region could be skipped. We check
if the any part of the surface is not hidden by the oapque region
above the surface.

If the surface is found to be visible, we add its opaque region to
the opaque reguion that may hide the underlying surfaces.