v4l2-renderer: Pass an opaque region of a view to render.
When V4L2 renderer calls the underlying renderer device, it now
passes a view's opaque region as well.

The original algorithm for calculating opaque region introduced
in commit ID 22948e8019fab52f73bc5f61dfeaf0ecd5836505 had an issue.
This is also fixed by this change.

The entire surface region is handled in the global coordinate. This
is translated to the output coordinate system, and then we figure out
which part needs to be rendered in the surface local coordinate system,
i.e we figure out clip region in the surface local coordinate.

However, the opaque region is other way around. We know the region in
the surface local coordinate. We have to translate this into the output
coordinate system, intersect with the output region, and then inverse
translate to figure out the clip region in the surface local coordinate.

Naturally, to figure out the opaque region for the renering optimization
we should use the one in the output coordinate system, which we didn't.