vsp-renderer: gl-fallback: make a number of views to compose with VSP configurable.
Make a number of views to compose with VSP configurable via weston.ini.
The number of views to be compose with VSP can be limited by putting the
folowing lines in the weston configuration file:

[vsp-renderer]
max_views_to_compose=4

This limits the number of views to be composed to 4 views at most, and
falls back to other renderer if possible.