v4l2-renderer: Don't release dmabuf buffer immediately when client release buffer
Since a dmabuf buffer was released immediately when a client released
buffer, surface flickering was occured when the surface is resized and
fade animation didn't run when client finished. By having a reference of
attached buffer until other dmabuf buffer is attached or the surface is
destroyed, these issues are resolved.