v4l2-renderer: Release dma-buf when attaching null buffer
When renderer is requested to attach a null buffer, e.g. via
wl_surface_attach with wl_buffer pointer set to null, it is
expected that the current buffer is released, effectively
dropping our reference on the attached dma buffer. Otherwise,
the reference is kept until the renderer is destroyed preventing
the dma buffer from being released earlier.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>